requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.4
django-appconf==1.0.4
django-compressor==2.4
django-filter==2.2.0
django-filters==0.2.1
django-libsass==0.8
djangorestframework==3.11.0
gunicorn==20.0.4
libsass==0.19.4
psycopg2-binary==2.8.4
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
